refactor: Adding logging module for log in and log out functions
