+++ app/controllers/users_controller.rb
@@ -1,4 +1,6 @@
+require 'logging_module'
 class UsersController < ApplicationController
+  include LoggingModule
   def user
     @user = User.new
   end
