+++ app/javascript/components/hello_name.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
 
 function HelloName (props) {
