+++ lib/logging_module.rb
@@ -1,0 +1,26 @@
+module LoggingModule
+  def login
+    session[:user_id] = @user.id
+  end
+
+  def logout
+    session.clear
+  end
+
+  def logged_in?
+    !!session[:user_id]
+  end
+
+  def current_user
+    @current_user ||= User.find(session[:user_id]) if session[:user_id]
+  end
+
+  def user_name?
+    @user = current_user
+    @user.first_name ||= User.find(session[:user_id]) if session[:user_id]
+  end
+
+  def set_user
+    @user = User.find_by(id: session[:user_id])
+  end
+end
